Revert "[web] Implement SceneBuilder.pushColorFilter for html (#20802)" (#20820)
This reverts commit 2decf195d2def7aa82b2d6e4c29ce932a78c73f2.
