+++ lib/web_ui/dev/goldens_lock.yaml
@@ -1,2 +1,2 @@
 repository: https://github.com/flutter/goldens.git
-revision: 8a831253654d151635ee9cfb71389c257413de5d
+revision: f26d68c3596eece3d40112e9dff01dc55d9bae97
