+++ lib/web_ui/lib/src/engine/bitmap_canvas.dart
@@ -573,10 +573,57 @@ class BitmapCanvas extends EngineCanvas {
       ui.Color? filterColor, ui.BlendMode colorFilterBlendMode,
       SurfacePaintData paint) {
     // For srcIn blendMode, we use an svg filter to apply to image element.
-    String? svgFilter = svgFilterFromBlendMode(filterColor,
-        colorFilterBlendMode);
+    String? svgFilter;
+    switch (colorFilterBlendMode) {
+      case ui.BlendMode.srcIn:
+      case ui.BlendMode.srcATop:
+        svgFilter = _srcInColorFilterToSvg(filterColor);
+        break;
+      case ui.BlendMode.srcOut:
+        svgFilter = _srcOutColorFilterToSvg(filterColor);
+        break;
+      case ui.BlendMode.xor:
+        svgFilter = _xorColorFilterToSvg(filterColor);
+        break;
+      case ui.BlendMode.plus:
+        // Porter duff source + destination.
+        svgFilter = _compositeColorFilterToSvg(filterColor, 0, 1, 1, 0);
+        break;
+      case ui.BlendMode.modulate:
+        // Porter duff source * destination but preserves alpha.
+        svgFilter = _modulateColorFilterToSvg(filterColor!);
+        break;
+      case ui.BlendMode.overlay:
+        // Since overlay is the same as hard-light by swapping layers,
+        // pass hard-light blend function.
+        svgFilter = _blendColorFilterToSvg(filterColor, 'hard-light',
+            swapLayers: true);
+        break;
+      // Several of the filters below (although supported) do not render the
+      // same (close but not exact) as native flutter when used as blend mode
+      // for a background-image with a background color. They only look
+      // identical when feBlend is used within an svg filter definition.
+      //
+      // Saturation filter uses destination when source is transparent.
+      // cMax = math.max(r, math.max(b, g));
+      // cMin = math.min(r, math.min(b, g));
+      // delta = cMax - cMin;
+      // lightness = (cMax + cMin) / 2.0;
+      // saturation = delta / (1.0 - (2 * lightness - 1.0).abs());
+      case ui.BlendMode.saturation:
+      case ui.BlendMode.colorDodge:
+      case ui.BlendMode.colorBurn:
+      case ui.BlendMode.hue:
+      case ui.BlendMode.color:
+      case ui.BlendMode.luminosity:
+        svgFilter = _blendColorFilterToSvg(filterColor,
+            _stringForBlendMode(colorFilterBlendMode));
+        break;
+      default:
+        break;
+    }
     final html.Element filterElement =
-        html.Element.html(svgFilter, treeSanitizer: _NullTreeSanitizer());
+    html.Element.html(svgFilter, treeSanitizer: _NullTreeSanitizer());
     rootElement.append(filterElement);
     _children.add(filterElement);
     final html.HtmlElement imgElement = _reuseOrCreateImage(image);
@@ -966,3 +1013,109 @@ String _maskFilterToCanvasFilter(ui.MaskFilter? maskFilter) {
   }
 }
 
+int _filterIdCounter = 0;
+
+// The color matrix for feColorMatrix element changes colors based on
+// the following:
+//
+// | R' |     | r1 r2 r3 r4 r5 |   | R |
+// | G' |     | g1 g2 g3 g4 g5 |   | G |
+// | B' |  =  | b1 b2 b3 b4 b5 | * | B |
+// | A' |     | a1 a2 a3 a4 a5 |   | A |
+// | 1  |     | 0  0  0  0  1  |   | 1 |
+//
+// R' = r1*R + r2*G + r3*B + r4*A + r5
+// G' = g1*R + g2*G + g3*B + g4*A + g5
+// B' = b1*R + b2*G + b3*B + b4*A + b5
+// A' = a1*R + a2*G + a3*B + a4*A + a5
+String _srcInColorFilterToSvg(ui.Color? color) {
+  _filterIdCounter += 1;
+  return '<svg width="0" height="0">'
+      '<filter id="_fcf$_filterIdCounter" '
+      'filterUnits="objectBoundingBox" x="0%" y="0%" width="100%" height="100%">'
+      '<feColorMatrix values="0 0 0 0 1 ' // Ignore input, set it to absolute.
+          '0 0 0 0 1 '
+          '0 0 0 0 1 '
+          '0 0 0 1 0" result="destalpha"/>' // Just take alpha channel of destination
+      '<feFlood flood-color="${colorToCssString(color)}" flood-opacity="1" result="flood">'
+      '</feFlood>'
+      '<feComposite in="flood" in2="destalpha" '
+      'operator="arithmetic" k1="1" k2="0" k3="0" k4="0" result="comp">'
+      '</feComposite>'
+      '</filter></svg>';
+}
+
+String _srcOutColorFilterToSvg(ui.Color? color) {
+  _filterIdCounter += 1;
+  return '<svg width="0" height="0">'
+      '<filter id="_fcf$_filterIdCounter" '
+      'filterUnits="objectBoundingBox" x="0%" y="0%" width="100%" height="100%">'
+      '<feFlood flood-color="${colorToCssString(color)}" flood-opacity="1" result="flood">'
+      '</feFlood>'
+      '<feComposite in="flood" in2="SourceGraphic" operator="out" result="comp">'
+      '</feComposite>'
+      '</filter></svg>';
+}
+
+String _xorColorFilterToSvg(ui.Color? color) {
+  _filterIdCounter += 1;
+  return '<svg width="0" height="0">'
+      '<filter id="_fcf$_filterIdCounter" '
+      'filterUnits="objectBoundingBox" x="0%" y="0%" width="100%" height="100%">'
+      '<feFlood flood-color="${colorToCssString(color)}" flood-opacity="1" result="flood">'
+      '</feFlood>'
+      '<feComposite in="flood" in2="SourceGraphic" operator="xor" result="comp">'
+      '</feComposite>'
+      '</filter></svg>';
+}
+
+// The source image and color are composited using :
+// result = k1 *in*in2 + k2*in + k3*in2 + k4.
+String _compositeColorFilterToSvg(ui.Color? color, double k1, double k2, double k3 , double k4) {
+  _filterIdCounter += 1;
+  return '<svg width="0" height="0">'
+      '<filter id="_fcf$_filterIdCounter" '
+      'filterUnits="objectBoundingBox" x="0%" y="0%" width="100%" height="100%">'
+      '<feFlood flood-color="${colorToCssString(color)}" flood-opacity="1" result="flood">'
+      '</feFlood>'
+      '<feComposite in="flood" in2="SourceGraphic" '
+      'operator="arithmetic" k1="$k1" k2="$k2" k3="$k3" k4="$k4" result="comp">'
+      '</feComposite>'
+      '</filter></svg>';
+}
+
+// Porter duff source * destination , keep source alpha.
+// First apply color filter to source to change it to [color], then
+// composite using multiplication.
+String _modulateColorFilterToSvg(ui.Color color) {
+  _filterIdCounter += 1;
+  final double r = color.red / 255.0;
+  final double b = color.blue / 255.0;
+  final double g = color.green / 255.0;
+  return '<svg width="0" height="0">'
+      '<filter id="_fcf$_filterIdCounter" '
+      'filterUnits="objectBoundingBox" x="0%" y="0%" width="100%" height="100%">'
+      '<feColorMatrix values="0 0 0 0 $r ' // Ignore input, set it to absolute.
+      '0 0 0 0 $g '
+      '0 0 0 0 $b '
+      '0 0 0 1 0" result="recolor"/>'
+      '<feComposite in="recolor" in2="SourceGraphic" '
+      'operator="arithmetic" k1="1" k2="0" k3="0" k4="0" result="comp">'
+      '</feComposite>'
+      '</filter></svg>';
+}
+
+// Uses feBlend element to blend source image with a color.
+String _blendColorFilterToSvg(ui.Color? color, String? feBlend,
+    {bool swapLayers = false}) {
+  _filterIdCounter += 1;
+  return '<svg width="0" height="0">'
+      '<filter id="_fcf$_filterIdCounter" filterUnits="objectBoundingBox" '
+      'x="0%" y="0%" width="100%" height="100%">'
+      '<feFlood flood-color="${colorToCssString(color)}" flood-opacity="1" result="flood">'
+      '</feFlood>' +
+      (swapLayers
+      ? '<feBlend in="SourceGraphic" in2="flood" mode="$feBlend"/>'
+      : '<feBlend in="flood" in2="SourceGraphic" mode="$feBlend"/>') +
+      '</filter></svg>';
+}
