--- lib/web_ui/test/golden_tests/engine/color_filter_golden_test.dart
@@ -1,89 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-// @dart = 2.6
-import 'dart:html' as html;
-
-import 'package:test/bootstrap/browser.dart';
-import 'package:test/test.dart';
-import 'package:ui/ui.dart';
-import 'package:ui/src/engine.dart';
-
-import 'package:web_engine_tester/golden_tester.dart';
-
-final Rect region = Rect.fromLTWH(0, 0, 500, 500);
-
-void main() {
-  internalBootstrapBrowserTest(() => testMain);
-}
-
-void testMain() async {
-  setUp(() async {
-    debugShowClipLayers = true;
-    SurfaceSceneBuilder.debugForgetFrameScene();
-    for (html.Node scene in html.document.querySelectorAll('flt-scene')) {
-      scene.remove();
-    }
-
-    await webOnlyInitializePlatform();
-    webOnlyFontCollection.debugRegisterTestFonts();
-    await webOnlyFontCollection.ensureFontsLoaded();
-  });
-
-  test('Should apply color filter to image', () async {
-    final SurfaceSceneBuilder builder = SurfaceSceneBuilder();
-    final Picture backgroundPicture = _drawBackground();
-    builder.addPicture(Offset.zero, backgroundPicture);
-    builder.pushColorFilter(EngineColorFilter.mode(Color(0xF0000080),
-        BlendMode.color));
-    final Picture circles1 = _drawTestPictureWithCircles(30, 30);
-    builder.addPicture(Offset.zero, circles1);
-    builder.pop();
-    html.document.body.append(builder
-        .build()
-        .webOnlyRootElement);
-
-    await matchGoldenFile('color_filter_blendMode_color.png', region: region);
-  });
-}
-
-Picture _drawTestPictureWithCircles(double offsetX, double offsetY) {
-  final EnginePictureRecorder recorder = PictureRecorder();
-  final RecordingCanvas canvas =
-  recorder.beginRecording(const Rect.fromLTRB(0, 0, 400, 400));
-  canvas.drawCircle(
-      Offset(offsetX + 10, offsetY + 10), 10, Paint()..style = PaintingStyle.fill);
-  canvas.drawCircle(
-      Offset(offsetX + 60, offsetY + 10),
-      10,
-      Paint()
-        ..style = PaintingStyle.fill
-        ..color = const Color.fromRGBO(255, 0, 0, 1));
-  canvas.drawCircle(
-      Offset(offsetX + 10, offsetY + 60),
-      10,
-      Paint()
-        ..style = PaintingStyle.fill
-        ..color = const Color.fromRGBO(0, 255, 0, 1));
-  canvas.drawCircle(
-      Offset(offsetX + 60, offsetY + 60),
-      10,
-      Paint()
-        ..style = PaintingStyle.fill
-        ..color = const Color.fromRGBO(0, 0, 255, 1));
-  return recorder.endRecording();
-}
-
-Picture _drawBackground() {
-  final EnginePictureRecorder recorder = PictureRecorder();
-  final RecordingCanvas canvas =
-  recorder.beginRecording(const Rect.fromLTRB(0, 0, 400, 400));
-  canvas.drawRect(
-      Rect.fromLTWH(8, 8, 400.0 - 16, 400.0 - 16),
-      Paint()
-        ..style = PaintingStyle.fill
-        ..color = Color(0xFFE0FFE0)
-  );
-  return recorder.endRecording();
-}
